change to Posts component
