+++ frontend/styles/indexStyles.tsx
